+++ src/pages/SignIn/SignIn.tsx
@@ -1,42 +1,16 @@
-import React, { useRef } from 'react'
+import React from 'react'
 
 import { MdAlternateEmail } from 'react-icons/md'
-
-import { useForm } from 'react-hook-form'
-import { yupResolver } from '@hookform/resolvers/yup'
-import * as yup from 'yup'
 
 import { Button, Input, PasswordInput } from 'ui'
 
 import { Layout } from './Layout'
 
+import { useSignIn } from './useSignIn'
 import * as S from './SignIn.styled'
 
-type UseForm = {
-  email: string
-  password: string
-}
-
-const schema = yup.object().shape({
-  email: yup.string().email().required(),
-  password: yup.string().required(),
-})
-
 export function SignIn() {
-  const buttonRef = useRef<HTMLButtonElement>(null)
-
-  const {
-    register,
-    handleSubmit,
-    setFocus,
-    formState: { errors },
-  } = useForm<UseForm>({
-    resolver: yupResolver(schema),
-  })
-
-  function handleOnSubmit() {
-    alert('Submited')
-  }
+  const { errors, register, setFocus, onSubmit } = useSignIn()
 
   return (
     <Layout>
@@ -46,7 +20,7 @@ export function SignIn() {
           Take full control of your financial life from
           <br /> now on with our help
         </h4>
-        <form onSubmit={handleSubmit(handleOnSubmit)}>
+        <form onSubmit={onSubmit}>
           <Input
             id='email'
             label='Email'
@@ -61,15 +35,12 @@ export function SignIn() {
             label='Password'
             placeholder='Secret password'
             error={errors.password?.message}
-            onKeyDown={() => buttonRef?.current?.click()}
             {...register('password')}
           />
           <S.ForgotPasswordLink to='/forgot_password'>
             Forgot your password?
           </S.ForgotPasswordLink>
-          <Button ref={buttonRef} type='submit'>
-            Join in account
-          </Button>
+          <Button type='submit'>Join in account</Button>
           <S.Separator>OR</S.Separator>
           <S.CreateAccountLink to='/register'>
             Create an account
